requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.3
python-dotenv==1.0.1
slowapi==0.1.8
argon2-cffi==23.1.0
pyotp==2.9.0
fido2==1.1.2
aiosmtplib==3.0.1
cryptography==42.0.5
httpx==0.27.0
pydantic[email]==1.10.14
anyio==4.3.0
pytest==8.1.1
pytest-asyncio==0.23.5
coverage==7.4.1
testcontainers[redis,postgres]==3.7.1
gunicorn==21.2.0
